Specify Package Version in Separate VERSION File
Following the guide on
https://packaging.python.org/guides/single-sourcing-package-version/
the LICENSE file has to be specified not only in the setup.cfg but
also included within the MANIFEST.in.
Also, the specified version has to comply with PEP 440.
To have the VERSION file also available from within the package
if it is installed with pip install -e ycbvideo, the VERSION
file has to be placed inside the package directory, not in the
root directory.
Additionally, to have the VERSION file available from within the
package if its .whl is installed, include_package_data = True
has to be set in setup.cfg.