Extend Tests
A README is added describing the correct test setup. Also, tests are
added to make sure that an exception is raised immediately when a
necessary file for a requested frame is missing, even if the frame
is not specified first by the corresponding selection expression.
An error regarding the test, where an exception is expected when
an invalid selection expression is provided, is fixed.