Add Frame Info
The frame info is shown when the package itself is executed with
python -m ycbvideo.
Using options.package_data in setup.cfg (documented in
https://docs.python.org/3.8/distutils/setupscript.html#distutils-installing-package-data)
to include the frame_info.txt does
not work, at least not for the wheel distribution. Therefore
the file is included in the build directly by adding it to
MANIFEST.in as done in the past for the VERSION file.