Let frames Return an Object Supporting len()
Since the number of frames selected is relevant for e.g.
the usage of the package together with Pytorch's
torch.utils.data.Dataset, where a __len__ method is expected,
frames() now returns an object, supporting the __len__ method
as well instead of "only" providing an iterator.
Furthermore, this objects supports the __iter__ method but no
__next__ method, therefore the type hint on the return type of
frames() is changed to "Iterable".
Since the for loop works basically the same on iterators and
iterables (an iterable internally returns an iterator when
__iter__ is called, an iterator returns itself when __iter__
is called), the object returned by frames() can be used exactly
as before.