Fix Problem When Parsing *-box.txt Files with Negative Values
Instead of matching the entries in a *-box.txt file with a regex
pattern as before, they are now simply expected to be correct
(i.e. a class label followed by four floating point numbers for
the four bounding box coordinates on each line), since the regex
used before did not match for negative numbers and extending the
regex for all possible cases would be of no great benefit.
Furthermore the bounding box coordinates are now already bundled
in two tuples, each containing the x- and y-coordinate of a corner
of the corresponding bounding box.