Fix Subtle Bugs Regarding Missing Files and data_syn
A frame, which misses a file but should be selected by a range
expression, where the frame is no "named" element, i.e. the frame
is neither specified as the start nor the stop in the range
expression, was silently ignored before, since only complete frames
were considered in the selection process. Only if the incomplete
frame was explicitly "named", an error was raised, since in this
case it was not found in the complete frames and therefore a
MissingElementError was raised. Now all available frames are
considered in the selection process. Afterwards, it is made sure
that now incomplete frames were selected, otherwise an appropriate
error is raised.
When a sequence range expression neither specified start nor stop
(i.e. ':') and data_syn available, data_syn was part of the
resulting selection. Now, the elements to select from are filtered
before the selection is made.